+++ src/xrEngine/FDemoPlay.cpp
@@ -51,7 +51,7 @@ CDemoPlay::CDemoPlay(const char *name, float ms, u32 cycles, float life_time) : 
 		}
 		
 		seq.resize		(sz/sizeof(Fmatrix));
-		m_count			= seq.size();
+		m_count			= (u32)seq.size();
 		CopyMemory	(&*seq.begin(),fs->pointer(),sz);
 		FS.r_close		(fs);
 		Log				("~ Total key-frames: ",m_count);
