+++ src/xrEngine/LightAnimLibrary.cpp
@@ -50,7 +50,7 @@ void CLAItem::Save(IWriter& F)
 	F.close_chunk	();
 
     F.open_chunk	(CHUNK_ITEM_KEYS);
-    F.w_u32		(Keys.size());
+    F.w_u32		((u32)Keys.size());
     for (KeyPairIt it=Keys.begin(); it!=Keys.end(); it++){
 		F.w_u32	(it->first);
 		F.w_u32	(it->second);
