+++ src/xrEngine/editor_environment_suns_flares.cpp
@@ -111,7 +111,7 @@ void flares::fill	(manager const& manager, editor::property_holder* holder, edit
 		m_shader.c_str(),
 		m_shader,
 		&*manager.m_environment.shader_ids().begin(),
-		manager.m_environment.shader_ids().size(),
+		(u32)manager.m_environment.shader_ids().size(),
 		editor::property_holder::value_editor_tree_view,
 		editor::property_holder::cannot_enter_text
 	);
