+++ src/xrGame/game_cl_base_weapon_usage_statistic_save.cpp
@@ -140,7 +140,7 @@ void Player_Statistic::WriteLtx(CInifile& ini, LPCSTR sect)
 
 	ini.w_u8(sect,"dwCurrentTeam",m_dwCurrentTeam);
 
-	u32 NumWeapons = aWeaponStats.size();
+	u32 NumWeapons = (u32)aWeaponStats.size();
 
 	ini.w_u32(sect,"NumWeapons",NumWeapons);
 
@@ -205,7 +205,7 @@ void Weapon_Statistic::WriteLtx(CInifile& ini, LPCSTR sect)
 
 	ini.w_u32(sect,"NumHits",NumHits);
 
-	u32 hits_size = m_Hits.size();
+	u32 hits_size = (u32)m_Hits.size();
 	i = 0;
 	u32 hit_number = 0;
 	u8 hit_index = 0;
